third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-script-preload-blocked-expected.txt

CONSOLE ERROR: line 6: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/require-sri-for/not-ran.js' because 'require-sri-for' directive requires integrity attribute be present for all scripts.
This is a testharness.js-based test.
PASS Script that requires integrity executes and does not generate a violation report. 
PASS Makes sure that require-sri-for applies to preloaded resources. 
Harness: the test ran to completion.
